fix: make sure custom richtext link fields are always part of a 'fields' group
