+++ src/admin/components/forms/field-types/RichText/elements/link/Element/index.tsx
@@ -3,7 +3,7 @@ import { ReactEditor, useSlate } from 'slate-react';
 import { Transforms, Node, Editor } from 'slate';
 import { useModal } from '@faceless-ui/modal';
 import { Trans, useTranslation } from 'react-i18next';
-import { unwrapLink } from '../utilities';
+import { transformExtraFields, unwrapLink } from '../utilities';
 import Popup from '../../../../../../elements/Popup';
 import { LinkDrawer } from '../LinkDrawer';
 import { Fields } from '../../../../../Form/types';
@@ -11,8 +11,6 @@ import buildStateFromSchema from '../../../../../Form/buildStateFromSchema';
 import { useAuth } from '../../../../../../utilities/Auth';
 import { useLocale } from '../../../../../../utilities/Locale';
 import { useConfig } from '../../../../../../utilities/Config';
-import { getBaseFields } from '../LinkDrawer/baseFields';
-import { Field } from '../../../../../../../../fields/config/types';
 import reduceFieldsToValues from '../../../../../Form/reduceFieldsToValues';
 import deepCopyObject from '../../../../../../../../utilities/deepCopyObject';
 import Button from '../../../../../../elements/Button';
@@ -23,6 +21,10 @@ import { useDrawerSlug } from '../../../../../../elements/Drawer/useDrawerSlug';
 
 const baseClass = 'rich-text-link';
 
+/**
+ * This function is called when an existing link is edited.
+ * When a link is first created, another function is called: {@link ../Button/index.tsx#insertLink}
+ */
 const insertChange = (editor, fields, customFieldSchema) => {
   const data = reduceFieldsToValues(fields, true);
 
@@ -79,25 +81,8 @@ export const LinkElement: React.FC<{
   const [renderPopup, setRenderPopup] = useState(false);
   const [initialState, setInitialState] = useState<Fields>({});
   const [fieldSchema] = useState(() => {
-    const baseFields: Field[] = getBaseFields(config);
+    const fields = transformExtraFields(customFieldSchema, config, i18n);
 
-    const fields = typeof customFieldSchema === 'function' ? customFieldSchema({ defaultFields: baseFields, config, i18n }) : baseFields;
-
-    if (Array.isArray(customFieldSchema)) {
-      fields.push({
-        name: 'fields',
-        type: 'group',
-        admin: {
-          style: {
-            margin: 0,
-            padding: 0,
-            borderTop: 0,
-            borderBottom: 0,
-          },
-        },
-        fields: customFieldSchema,
-      });
-    }
 
     return fields;
   });
